DN-12: Refactor project files.
