+++ client/package.json
@@ -9,6 +9,7 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
+    "@derecksnotes/shared": "workspace:*",
     "@tanstack/react-table": "^8.21.3",
     "axios": "^1.7.9",
     "date-fns": "^4.1.0",
