+++ client/src/lib/env.ts
@@ -1,29 +1,10 @@
 import pkg from '../../../package.json';
-
-type BuildEnv = 'local' | 'dev' | 'prod';
+import { ENV_CONFIG, type BuildEnv } from '@derecksnotes/shared';
 
 const BUILD_ENV =
     (process.env.NEXT_PUBLIC_BUILD_ENV as BuildEnv) ||
     (process.env.BUILD_ENV as BuildEnv) ||
     'local';
-
-const ENV_CONFIG = {
-    local: {
-        domain: 'localhost',
-        baseUrl: 'http://localhost:3000',
-        apiUrl: 'http://localhost:3000/api'
-    },
-    dev: {
-        domain: 'dev.derecksnotes.com',
-        baseUrl: 'https://dev.derecksnotes.com',
-        apiUrl: 'https://dev.derecksnotes.com/api'
-    },
-    prod: {
-        domain: 'derecksnotes.com',
-        baseUrl: 'https://derecksnotes.com',
-        apiUrl: 'https://derecksnotes.com/api'
-    }
-} as const;
 
 const derived = ENV_CONFIG[BUILD_ENV];
