+++ client/tsconfig.json
@@ -28,6 +28,12 @@
       ],
       "@public/*": [
         "./public/*"
+      ],
+      "@derecksnotes/shared": [
+        "../shared/src/index.ts"
+      ],
+      "@derecksnotes/shared/*": [
+        "../shared/src/*"
       ]
     }
   },
