+++ package.json
@@ -1,13 +1,16 @@
 {
   "name": "derecksnotes.com",
-  "version": "5.1.1",
+  "version": "6.0.0",
   "type": "module",
+  "workspaces": ["client", "server", "shared"],
   "repository": "git@github.com:dereckdemezquita/derecksnotes.com.git",
   "scripts": {
     "dev": "concurrently -n client,server -c blue,green \"bun run --cwd client dev\" \"bun run --cwd server dev\"",
     "build": "bun run --cwd client build",
     "start": "concurrently -n client,server \"bun run --cwd client start\" \"bun run --cwd server start\"",
     "format": "bun prettier --write .",
+    "lint": "bun run --cwd client lint",
+    "typecheck": "bun run --cwd client typecheck",
     "test": "bun test",
     "db:generate": "bun run --cwd server db:generate",
     "db:migrate": "bun run --cwd server db:migrate",
