+++ server/package.json
@@ -24,6 +24,7 @@
   },
   "type": "module",
   "dependencies": {
+    "@derecksnotes/shared": "workspace:*",
     "bcrypt": "^6.0.0",
     "cookie-parser": "^1.4.7",
     "cors": "^2.8.5",
