+++ server/tsconfig.json
@@ -34,7 +34,9 @@
       "@services/*": ["./src/services/*"],
       "@routes/*": ["./src/routes/*"],
       "@middleware/*": ["./src/middleware/*"],
-      "@lib/*": ["./src/lib/*"]
+      "@lib/*": ["./src/lib/*"],
+      "@derecksnotes/shared": ["../shared/src/index.ts"],
+      "@derecksnotes/shared/*": ["../shared/src/*"]
     }
   }
 }
