+++ shared/package.json
@@ -1,0 +1,14 @@
+{
+    "name": "@derecksnotes/shared",
+    "version": "1.0.0",
+    "private": true,
+    "type": "module",
+    "main": "src/index.ts",
+    "types": "src/index.ts",
+    "exports": {
+        ".": "./src/index.ts",
+        "./types": "./src/types/index.ts",
+        "./types/*": "./src/types/*",
+        "./env": "./src/env.ts"
+    }
+}
